+++ source/dem/particle_wall_contact_force.cc
@@ -3,203 +3,788 @@
 
 #include <dem/particle_wall_contact_force.h>
 
-// Updates the contact information (contact_info) based on the new information
-// of particles pair in the current time step
-template <int dim, typename PropertiesIndex>
-void
-ParticleWallContactForce<dim, PropertiesIndex>::update_contact_information(
-  particle_wall_contact_info<dim> &contact_info,
-  const Point<3>                  &particle_position,
-  const ArrayView<const double>   &particle_properties,
-  const double                     dt)
+using namespace DEM;
+using namespace Parameters::Lagrangian;
+
+template <int dim,
+          typename PropertiesIndex,
+          ParticleWallContactForceModel contact_model,
+          RollingResistanceMethod       rolling_friction_model>
+ParticleWallContactForce<dim,
+                         PropertiesIndex,
+                         contact_model,
+                         rolling_friction_model>::
+  ParticleWallContactForce(
+    const DEMSolverParameters<dim>        &dem_parameters,
+    const std::vector<types::boundary_id> &boundary_index)
+  : dmt_cut_off_threshold(dem_parameters.model_parameters.dmt_cut_off_threshold)
+  , f_coefficient_epsd(dem_parameters.model_parameters.f_coefficient_epsd)
 {
-  // i is the particle, j is the wall.
-  // we need to put a minus sign infront of the normal_vector to respect the
-  // convention (i -> j)
-  auto               normal_vector = -contact_info.normal_vector;
-  const unsigned int boundary_id   = contact_info.boundary_id;
-
-  // Using velocity and angular velocity of particle as
-  // local vectors
-  Tensor<1, 3> particle_velocity;
-  particle_velocity[0] = particle_properties[PropertiesIndex::v_x];
-  particle_velocity[1] = particle_properties[PropertiesIndex::v_y];
-  particle_velocity[2] = particle_properties[PropertiesIndex::v_z];
-
-  Tensor<1, 3> particle_angular_velocity;
-  particle_angular_velocity[0] = particle_properties[PropertiesIndex::omega_x];
-  particle_angular_velocity[1] = particle_properties[PropertiesIndex::omega_y];
-  particle_angular_velocity[2] = particle_properties[PropertiesIndex::omega_z];
-
-  // Calculate approximation of the contact point using the normal vector
-  Point<3> contact_point =
-    particle_position +
-    0.5 * particle_properties[PropertiesIndex::dp] * normal_vector;
-
-  // Get vector pointing from the contact point to the origin of the rotation
-  // axis
-  Tensor<1, 3> vector_to_rotating_axis =
-    contact_point - this->point_on_rotation_vector[boundary_id];
-
-  // Remove the rotating axis component of that vector
-  vector_to_rotating_axis =
-    vector_to_rotating_axis -
-    (vector_to_rotating_axis * this->boundary_rotational_vector[boundary_id]) *
-      this->boundary_rotational_vector[boundary_id];
-
-  // Tensor<1,3> vector_to_rotation_axis = this->boundary_rotational_speed_map
-
-  // Defining relative contact velocity using the convention
-  // v_ij = v_j - v_i
-  Tensor<1, 3> contact_relative_velocity =
-    this->boundary_translational_velocity_map[boundary_id] - particle_velocity +
-    cross_product_3d((-0.5 * particle_properties[PropertiesIndex::dp] *
-                      particle_angular_velocity),
-                     normal_vector) +
-    cross_product_3d(this->boundary_rotational_speed_map[boundary_id] *
-                       this->boundary_rotational_vector[boundary_id],
-                     vector_to_rotating_axis);
-
-  // Calculation of normal relative velocity
-  double normal_relative_velocity_value =
-    contact_relative_velocity * normal_vector;
-  Tensor<1, 3> normal_relative_velocity =
-    normal_relative_velocity_value * normal_vector;
-
-  // Calculation of tangential relative velocity
-  Tensor<1, 3> tangential_relative_velocity =
-    contact_relative_velocity - normal_relative_velocity;
-
-  // Calculation of new tangential_displacement, since this value is
-  // history-dependent it needs the value at previous time-step
-  // This variable is the main reason that we have iteration over
-  // two different vectors (pairs_in_contact and
-  // contact_pair_candidates): tangential_displacement of the particles
-  // which were already in contact (pairs_in_contact) needs to be
-  // modified using its history, while the tangential_displacements of
-  // new particles are equal to zero
-  Tensor<1, 3> modified_tangential_displacement =
-    contact_info.tangential_displacement + tangential_relative_velocity * dt;
-
-  // Updating the contact_info container based on the new calculated values
-  contact_info.normal_relative_velocity     = normal_relative_velocity_value;
-  contact_info.tangential_displacement      = modified_tangential_displacement;
-  contact_info.tangential_relative_velocity = tangential_relative_velocity;
+  set_effective_properties(dem_parameters, boundary_index);
 }
 
-template <int dim, typename PropertiesIndex>
+template <int dim,
+          typename PropertiesIndex,
+          ParticleWallContactForceModel contact_model,
+          RollingResistanceMethod       rolling_friction_model>
 void
-ParticleWallContactForce<dim, PropertiesIndex>::
-  update_particle_floating_wall_contact_information(
-    particle_wall_contact_info<dim> &contact_info,
-    const ArrayView<const double>   &particle_properties,
-    const double                     dt,
-    const Tensor<1, 3>              &cut_cell_translational_velocity,
-    const Tensor<1, 3>              &cut_cell_rotational_velocity,
-    const double                     center_of_rotation_particle_distance)
+ParticleWallContactForce<dim,
+                         PropertiesIndex,
+                         contact_model,
+                         rolling_friction_model>::
+  calculate_particle_wall_contact_force(
+    typename DEM::dem_data_structures<dim>::particle_wall_in_contact
+                &particle_wall_pairs_in_contact,
+    const double dt,
+    ParticleInteractionOutcomes<PropertiesIndex> &contact_outcome)
 {
-  // i is the particle, j is the wall.
-  // we need to put a minus sign infront of the normal_vector to respect the
-  // convention (i -> j)
-  const Tensor<1, 3> normal_vector = -contact_info.normal_vector;
-
-  // Using velocity and angular velocity of particle as
-  // local vectors
-  Tensor<1, 3> particle_velocity;
-  particle_velocity[0] = particle_properties[PropertiesIndex::v_x];
-  particle_velocity[1] = particle_properties[PropertiesIndex::v_y];
-  particle_velocity[2] = particle_properties[PropertiesIndex::v_z];
-
-
-  Tensor<1, 3> particle_angular_velocity;
-  particle_angular_velocity[0] = particle_properties[PropertiesIndex::omega_x];
-  particle_angular_velocity[1] = particle_properties[PropertiesIndex::omega_y];
-  particle_angular_velocity[2] = particle_properties[PropertiesIndex::omega_z];
-
-  // Defining relative contact velocity
-  // v_ij = v_j - v_i
-  Tensor<1, 3> contact_relative_velocity =
-    cut_cell_translational_velocity - particle_velocity +
-    cross_product_3d((center_of_rotation_particle_distance *
-                        cut_cell_rotational_velocity -
-                      0.5 * particle_properties[PropertiesIndex::dp] *
-                        particle_angular_velocity),
-                     normal_vector);
-
-  // Calculation of normal relative velocity
-  double normal_relative_velocity_value =
-    contact_relative_velocity * normal_vector;
-  Tensor<1, 3> normal_relative_velocity =
-    normal_relative_velocity_value * normal_vector;
-
-  // Calculation of tangential relative velocity
-  Tensor<1, 3> tangential_relative_velocity =
-    contact_relative_velocity - normal_relative_velocity;
-
-  // Calculate the new tangential_displacement, since this value is
-  // history-dependent, it needs the value at previous time-step
-  Tensor<1, 3> modified_tangential_displacement =
-    contact_info.tangential_displacement + tangential_relative_velocity * dt;
-
-  // Updating the contact_info container based on the new calculated values
-  contact_info.normal_relative_velocity     = normal_relative_velocity_value;
-  contact_info.tangential_displacement      = modified_tangential_displacement;
-  contact_info.tangential_relative_velocity = tangential_relative_velocity;
-}
-
-template <int dim, typename PropertiesIndex>
-void
-ParticleWallContactForce<dim, PropertiesIndex>::
-  calculate_force_and_torque_on_boundary(const unsigned int boundary_id,
-                                         Tensor<1, 3>       add_force,
-                                         const Point<3>     point_contact)
-{
-  if (calculate_force_torque_on_boundary)
+  // Looping over particle_wall_pairs_in_contact, which means looping over
+  // all the active particles with iterator
+  // particle_wall_pairs_in_contact_iterator
+  for (auto &&pairs_in_contact_content :
+       particle_wall_pairs_in_contact | boost::adaptors::map_values)
     {
-      mpi_correction_over_calculation_of_forces_and_torques();
+      // Now an iterator (particle_wall_contact_info_iterator) on each
+      // element of the particle_wall_pairs_in_contact vector is defined.
+      // This iterator iterates over a map which contains the required
+      // information for calculation of the contact force for each particle
+      for (auto &&contact_info :
+           pairs_in_contact_content | boost::adaptors::map_values)
+        {
+          // Defining contact forces and properties of particle as
+          // local variables
+          auto particle            = contact_info.particle;
+          auto particle_properties = particle->get_properties();
 
-      force_on_walls[boundary_id] = force_on_walls[boundary_id] - add_force;
+          Tensor<1, 3> normal_vector     = contact_info.normal_vector;
+          auto         point_on_boundary = contact_info.point_on_boundary;
+          Tensor<1, 3> normal_force;
+          Tensor<1, 3> tangential_force;
+          Tensor<1, 3> tangential_torque;
+          Tensor<1, 3> rolling_resistance_torque;
 
-      torque_on_walls[boundary_id] =
-        torque_on_walls[boundary_id] -
-        cross_product_3d(point_contact - center_mass_container, add_force);
+          // Get particle 3d location
+          Point<3> particle_location_3d = get_location(particle);
+
+          // A vector (point_to_particle_vector) is defined which connects
+          // the center of particle to the point_on_boundary. This vector
+          // will then be projected on the normal vector of the boundary to
+          // obtain the particle-wall distance
+          Tensor<1, 3> point_to_particle_vector =
+            particle_location_3d - point_on_boundary;
+
+          // Finding the projected vector on the normal vector of the
+          // boundary. Here we have used the private function
+          // find_projection. Using this projected vector, the particle-wall
+          // distance is calculated
+          Tensor<1, 3> projected_vector =
+            this->find_projection(point_to_particle_vector, normal_vector);
+          double normal_overlap =
+            ((particle_properties[PropertiesIndex::dp]) * 0.5) -
+            (projected_vector.norm());
+
+          // Get the threshold distance for contact force, this is useful
+          // for non-contact cohesive force models such as the DMT.
+          const double force_calculation_threshold_distance =
+            get_force_calculation_threshold_distance();
+
+          if (normal_overlap > force_calculation_threshold_distance)
+            {
+              contact_info.normal_overlap = normal_overlap;
+
+              // Update all the information
+              this->update_contact_information(contact_info,
+                                               particle_location_3d,
+                                               particle_properties,
+                                               dt);
+
+              // Calculating the contact force
+              this->calculate_contact(contact_info,
+                                      dt,
+                                      particle_properties,
+                                      normal_force,
+                                      tangential_force,
+                                      tangential_torque,
+                                      rolling_resistance_torque);
+
+              // Apply the calculated forces and torques on the particle
+              types::particle_index particle_id = particle->get_local_index();
+              Tensor<1, 3>         &particle_torque =
+                contact_outcome.torque[particle_id];
+              Tensor<1, 3> &particle_force = contact_outcome.force[particle_id];
+
+              this->apply_force_and_torque(normal_force,
+                                           tangential_force,
+                                           tangential_torque,
+                                           rolling_resistance_torque,
+                                           particle_torque,
+                                           particle_force,
+                                           point_on_boundary,
+                                           contact_info.boundary_id);
+            }
+          else
+            {
+              contact_info.normal_overlap = 0.;
+              contact_info.tangential_displacement.clear();
+              contact_info.rolling_resistance_spring_torque.clear();
+            }
+        }
     }
 }
 
-template <int dim, typename PropertiesIndex>
-std::map<unsigned int, Tensor<1, 3>>
-ParticleWallContactForce<dim, PropertiesIndex>::initialize()
-{
-  std::map<unsigned int, Tensor<1, 3>> map;
-  for (const auto &it : boundary_index)
-    {
-      map[it] = 0;
-    }
-  return map;
-}
-
-template <int dim, typename PropertiesIndex>
+template <int dim,
+          typename PropertiesIndex,
+          ParticleWallContactForceModel contact_model,
+          RollingResistanceMethod       rolling_friction_model>
 void
-ParticleWallContactForce<dim, PropertiesIndex>::
-  mpi_correction_over_calculation_of_forces_and_torques()
+ParticleWallContactForce<dim,
+                         PropertiesIndex,
+                         contact_model,
+                         rolling_friction_model>::
+  calculate_particle_solid_object_contact_force(
+    typename DEM::dem_data_structures<dim>::particle_floating_mesh_in_contact
+                &particle_floating_mesh_in_contact,
+    const double dt,
+    const std::vector<std::shared_ptr<SerialSolid<dim - 1, dim>>> &solids,
+    ParticleInteractionOutcomes<PropertiesIndex> &contact_outcome)
 {
-  for (const auto &it : boundary_index)
+  std::vector<Particles::ParticleIterator<dim>> particle_locations;
+  std::vector<Point<dim>> triangle(this->vertices_per_triangle);
+
+  for (unsigned int solid_counter = 0; solid_counter < solids.size();
+       ++solid_counter)
     {
-      force_on_walls[it] =
-        Utilities::MPI::sum(force_on_walls[it], MPI_COMM_WORLD);
-      torque_on_walls[it] =
-        Utilities::MPI::sum(torque_on_walls[it], MPI_COMM_WORLD);
+      // Get translational and rotational velocities and center of
+      // rotation of solid object
+      Tensor<1, 3> translational_velocity =
+        solids[solid_counter]->get_translational_velocity();
+      Tensor<1, 3> angular_velocity =
+        solids[solid_counter]->get_angular_velocity();
+      Point<3> center_of_rotation =
+        solids[solid_counter]->get_center_of_rotation();
+
+      auto &particle_floating_mesh_contact_pair =
+        particle_floating_mesh_in_contact[solid_counter];
+
+      for (auto &[cut_cell, map_info] : particle_floating_mesh_contact_pair)
+        {
+          if (!map_info.empty())
+            {
+              // Clear the particle locations vector for the new cut cell
+              particle_locations.clear();
+              const unsigned int n_particles = map_info.size();
+
+              // Gather all the particles locations in a vector
+              for (auto &&contact_info : map_info | boost::adaptors::map_values)
+                {
+                  particle_locations.push_back(contact_info.particle);
+                }
+
+              // Build triangle vector
+              for (unsigned int vertex = 0;
+                   vertex < this->vertices_per_triangle;
+                   ++vertex)
+                {
+                  // Find vertex-floating wall distance
+                  triangle[vertex] = cut_cell->vertex(vertex);
+                }
+
+              // Call find_particle_triangle_projection to get the
+              // projection of particles on the triangle
+              // (floating mesh cell)
+              auto particle_triangle_information = LetheGridTools::
+                find_particle_triangle_projection<dim, PropertiesIndex>(
+                  triangle, particle_locations, n_particles);
+
+              const std::vector<bool> pass_distance_check =
+                std::get<0>(particle_triangle_information);
+              const std::vector<Point<3>> projection_points =
+                std::get<1>(particle_triangle_information);
+              const std::vector<Tensor<1, 3>> normal_vectors =
+                std::get<2>(particle_triangle_information);
+
+              unsigned int particle_counter = 0;
+
+              for (auto &&contact_info : map_info | boost::adaptors::map_values)
+                {
+                  // If particle passes the distance check
+                  if (pass_distance_check[particle_counter])
+                    {
+                      // Define the total force of contact, properties of
+                      // particle as local parameters
+                      auto &particle            = contact_info.particle;
+                      auto  particle_properties = particle->get_properties();
+                      Tensor<1, 3> normal_force;
+                      Tensor<1, 3> tangential_force;
+                      Tensor<1, 3> tangential_torque;
+                      Tensor<1, 3> rolling_resistance_torque;
+
+                      const Point<3> &projection_point =
+                        projection_points[particle_counter];
+
+                      Point<3> particle_location_3d = get_location(particle);
+
+                      const double particle_triangle_distance =
+                        particle_location_3d.distance(projection_point);
+
+                      // Find normal overlap
+                      double normal_overlap =
+                        ((particle_properties[PropertiesIndex::dp]) * 0.5) -
+                        particle_triangle_distance;
+
+                      // Get the threshold distance for contact force, this is
+                      // useful for non-contact cohesive force models such as
+                      // the DMT.
+                      const double force_calculation_threshold_distance =
+                        get_force_calculation_threshold_distance();
+
+                      if (normal_overlap > force_calculation_threshold_distance)
+                        {
+                          contact_info.normal_overlap = normal_overlap;
+                          contact_info.normal_vector =
+                            normal_vectors[particle_counter];
+
+                          contact_info.point_on_boundary = projection_point;
+
+                          contact_info.boundary_id = solid_counter;
+
+                          this
+                            ->update_particle_solid_object_contact_information(
+                              contact_info,
+                              particle_properties,
+                              dt,
+                              translational_velocity,
+                              angular_velocity,
+                              center_of_rotation.distance(
+                                particle_location_3d));
+
+                          // Calculating the contact force
+                          this->calculate_contact(contact_info,
+                                                  dt,
+                                                  particle_properties,
+                                                  normal_force,
+                                                  tangential_force,
+                                                  tangential_torque,
+                                                  rolling_resistance_torque);
+
+                          // Apply the calculated forces and torques on the
+                          // particle
+                          types::particle_index particle_id =
+                            particle->get_local_index();
+                          Tensor<1, 3> &particle_torque =
+                            contact_outcome.torque[particle_id];
+                          Tensor<1, 3> &particle_force =
+                            contact_outcome.force[particle_id];
+
+                          this->apply_force_and_torque(
+                            normal_force,
+                            tangential_force,
+                            tangential_torque,
+                            rolling_resistance_torque,
+                            particle_torque,
+                            particle_force,
+                            projection_point,
+                            contact_info.boundary_id);
+                        }
+                      else
+                        {
+                          contact_info.normal_overlap = 0;
+                          contact_info.tangential_displacement.clear();
+                          contact_info.rolling_resistance_spring_torque.clear();
+                        }
+                    }
+                }
+              particle_counter++;
+            }
+        }
     }
 }
 
-
-template class ParticleWallContactForce<2, DEM::DEMProperties::PropertiesIndex>;
+// dem
+// No resistance
 template class ParticleWallContactForce<2,
-                                        DEM::CFDDEMProperties::PropertiesIndex>;
+                                        DEM::DEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::DMT,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<3,
+                                        DEM::DEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::DMT,
+                                        RollingResistanceMethod::no_resistance>;
 template class ParticleWallContactForce<2,
-                                        DEM::DEMMPProperties::PropertiesIndex>;
-template class ParticleWallContactForce<3, DEM::DEMProperties::PropertiesIndex>;
+                                        DEM::DEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::JKR,
+                                        RollingResistanceMethod::no_resistance>;
 template class ParticleWallContactForce<3,
-                                        DEM::CFDDEMProperties::PropertiesIndex>;
+                                        DEM::DEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::JKR,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<2,
+                                        DEM::DEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::linear,
+                                        RollingResistanceMethod::no_resistance>;
 template class ParticleWallContactForce<3,
-                                        DEM::DEMMPProperties::PropertiesIndex>;
+                                        DEM::DEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::linear,
+                                        RollingResistanceMethod::no_resistance>;
+
+// Constant resistance
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::constant_resistance>;
+
+// Viscous resistance
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::viscous_resistance>;
+
+// EPSD resistance
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::epsd_resistance>;
+
+// cfd_dem
+// No resistance
+template class ParticleWallContactForce<2,
+                                        DEM::CFDDEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::DMT,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<3,
+                                        DEM::CFDDEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::DMT,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<2,
+                                        DEM::CFDDEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::JKR,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<3,
+                                        DEM::CFDDEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::JKR,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<2,
+                                        DEM::CFDDEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::linear,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<3,
+                                        DEM::CFDDEMProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::linear,
+                                        RollingResistanceMethod::no_resistance>;
+
+// Constant resistance
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::constant_resistance>;
+
+// Viscous resistance
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::viscous_resistance>;
+
+// EPSD resistance
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::CFDDEMProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::epsd_resistance>;
+
+// dem_mp
+//  No resistance
+template class ParticleWallContactForce<2,
+                                        DEM::DEMMPProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::DMT,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<3,
+                                        DEM::DEMMPProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::DMT,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<2,
+                                        DEM::DEMMPProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::JKR,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<3,
+                                        DEM::DEMMPProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::JKR,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<2,
+                                        DEM::DEMMPProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::linear,
+                                        RollingResistanceMethod::no_resistance>;
+template class ParticleWallContactForce<3,
+                                        DEM::DEMMPProperties::PropertiesIndex,
+                                        ParticleWallContactForceModel::linear,
+                                        RollingResistanceMethod::no_resistance>;
+
+// Constant resistance
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::constant_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::constant_resistance>;
+
+// Viscous resistance
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::viscous_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::viscous_resistance>;
+
+// EPSD resistance
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::DMT,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::JKR,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::nonlinear,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  2,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::epsd_resistance>;
+template class ParticleWallContactForce<
+  3,
+  DEM::DEMMPProperties::PropertiesIndex,
+  ParticleWallContactForceModel::linear,
+  RollingResistanceMethod::epsd_resistance>;
